+++ packages/@tinacms/app/appFiles/src/preview.tsx
@@ -47,6 +47,13 @@ export const Preview = (
           setActiveQuery(event.data)
         }
       })
+      window.addEventListener('message', (event: MessageEvent<PostMessage>) => {
+        if (event?.data?.type === 'isEditMode') {
+          props.iframeRef?.current?.contentWindow?.postMessage({
+            type: 'tina:editMode',
+          })
+        }
+      })
     }
   }, [props.iframeRef.current])
 
@@ -101,13 +108,6 @@ const QueryMachine = (props: {
 
   React.useEffect(() => {
     if (props.iframeRef.current) {
-      window.addEventListener('message', (event: MessageEvent<PostMessage>) => {
-        if (event?.data?.type === 'isEditMode') {
-          props.iframeRef?.current?.contentWindow?.postMessage({
-            type: 'tina:editMode',
-          })
-        }
-      })
       send({ type: 'IFRAME_MOUNTED', value: props.iframeRef.current })
       if (props.payload.type === 'open') {
         send({ type: 'ADD_QUERY', value: props.payload })