+++ packages/@tinacms/toolkit/src/packages/react-sidebar/components/ResizeHandle.tsx
@@ -75,12 +75,12 @@ export const ResizeHandle = () => {
   return (
     <div
       onMouseDown={handleresizingSidebar}
-      className={`z-100 absolute top-1/2 right-0 w-2 h-32 bg-white rounded-r-md border border-gray-100 shadow-sm hover:shadow-md transition-all duration-150 ease-out transform translate-x-full -translate-y-1/2 group hover:bg-gray-50 ${
+      className={`z-100 absolute top-1/2 right-px w-2 h-32 bg-white rounded-r-md border border-gray-150 shadow-sm hover:shadow-md origin-left transition-all duration-150 ease-out transform translate-x-full -translate-y-1/2 group hover:bg-gray-50 ${
         displayState !== 'closed' ? `opacity-100` : `opacity-0`
-      } ${resizingSidebar ? `scale-110` : ``}`}
+      } ${resizingSidebar ? `scale-110` : `scale-90 hover:scale-100`}`}
       style={{ cursor: 'grab' }}
     >
-      <span className="absolute top-1/2 left-1/2 h-4/6 w-px bg-gray-300 transform -translate-y-1/2 -translate-x-1/2 opacity-50 transition-opacity duration-150 ease-out group-hover:opacity-100"></span>
+      <span className="absolute top-1/2 left-1/2 h-4/6 w-px bg-gray-200 transform -translate-y-1/2 -translate-x-1/2 opacity-30 transition-opacity duration-150 ease-out group-hover:opacity-100"></span>
     </div>
   )
 }