+++ examples/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,21 @@
 # e2e-next
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [7495f032b]
+- Updated dependencies [64599e300]
+- Updated dependencies [de693ea17]
+  - tinacms@1.1.1
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [11b32f712]
+  - tinacms@1.1.0
+
 ## 0.1.9
 
 ### Patch Changes