+++ packages/@tinacms/app/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tinacms/app
 
+## 1.0.3
+
+### Patch Changes
+
+- 66ed1452b: Updated so that `useEditState` does not require `useTina` to be called first.
+
+## 1.0.2
+
+### Patch Changes
+
+- f13878798: Fix a bug where meta data was added to the form values payload
+- c62d2bad4: Fix issue where tailwind styles from within the user config weren't being picked up
+
 ## 1.0.1
 
 ### Patch Changes